+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-whatsapp-cta-url-button.tsx
@@ -1,4 +1,4 @@
-import { WhatsappCTAUrlButton } from '@botonic/react'
+import { WhatsappCTAUrlButton, WhatsappCTAUrlHeaderType } from '@botonic/react'
 import React from 'react'
 
 import { FlowBuilderApi } from '../api'
@@ -10,6 +10,10 @@ export class FlowWhatsappCtaUrlButtonNode extends ContentFieldsBase {
   public code = ''
   public text = ''
   public header?: string
+  public headerType?: WhatsappCTAUrlHeaderType
+  public headerImage?: string
+  public headerVideo?: string
+  public headerDocument?: string
   public footer?: string
   public displayText = ''
   public url: string = ''
@@ -25,9 +29,10 @@ export class FlowWhatsappCtaUrlButtonNode extends ContentFieldsBase {
       locale,
       component.content.text
     )
-    whatsappCtaUrlButton.header = this.getTextByLocale(
-      locale,
-      component.content.header
+    FlowWhatsappCtaUrlButtonNode.setHeader(
+      whatsappCtaUrlButton,
+      component,
+      locale
     )
     whatsappCtaUrlButton.footer = this.getTextByLocale(
       locale,
@@ -47,12 +52,127 @@ export class FlowWhatsappCtaUrlButtonNode extends ContentFieldsBase {
     return whatsappCtaUrlButton
   }
 
+  private static setHeader(
+    whatsappCtaUrlButton: FlowWhatsappCtaUrlButtonNode,
+    component: HtWhatsappCTAUrlButtonNode,
+    locale: string
+  ) {
+    // If header is set, set header_type to text, this is need to be compatible with the old version of the component
+    if (component.content.header.length > 0) {
+      component.content.header_type = WhatsappCTAUrlHeaderType.Text
+    }
+
+    whatsappCtaUrlButton.headerType = component.content
+      .header_type as WhatsappCTAUrlHeaderType
+
+    if (component.content.header_type === WhatsappCTAUrlHeaderType.Text) {
+      whatsappCtaUrlButton.header = this.getTextByLocale(
+        locale,
+        component.content.header
+      )
+    }
+
+    if (
+      component.content.header_type === WhatsappCTAUrlHeaderType.Image &&
+      component.content.header_image
+    ) {
+      whatsappCtaUrlButton.headerImage = this.getAssetByLocale(
+        locale,
+        component.content.header_image
+      )
+    }
+
+    if (
+      component.content.header_type === WhatsappCTAUrlHeaderType.Video &&
+      component.content.header_video
+    ) {
+      whatsappCtaUrlButton.headerVideo = this.getAssetByLocale(
+        locale,
+        component.content.header_video
+      )
+    }
+
+    if (
+      component.content.header_type === WhatsappCTAUrlHeaderType.Document &&
+      component.content.header_document
+    ) {
+      whatsappCtaUrlButton.headerDocument = this.getAssetByLocale(
+        locale,
+        component.content.header_document
+      )
+    }
+  }
+
   toBotonic(id: string): JSX.Element {
+    if (
+      this.headerType === WhatsappCTAUrlHeaderType.Image &&
+      this.headerImage
+    ) {
+      return (
+        <WhatsappCTAUrlButton
+          key={id}
+          body={this.text}
+          headerType={this.headerType}
+          headerImage={this.headerImage}
+          footer={this.footer}
+          displayText={this.displayText}
+          url={this.url}
+        />
+      )
+    }
+
+    if (
+      this.headerType === WhatsappCTAUrlHeaderType.Video &&
+      this.headerVideo
+    ) {
+      return (
+        <WhatsappCTAUrlButton
+          key={id}
+          body={this.text}
+          headerType={this.headerType}
+          headerVideo={this.headerVideo}
+          footer={this.footer}
+          displayText={this.displayText}
+          url={this.url}
+        />
+      )
+    }
+
+    if (
+      this.headerType === WhatsappCTAUrlHeaderType.Document &&
+      this.headerDocument
+    ) {
+      return (
+        <WhatsappCTAUrlButton
+          key={id}
+          body={this.text}
+          headerType={this.headerType}
+          headerDocument={this.headerDocument}
+          footer={this.footer}
+          displayText={this.displayText}
+          url={this.url}
+        />
+      )
+    }
+
+    if (this.headerType === WhatsappCTAUrlHeaderType.Text && this.header) {
+      return (
+        <WhatsappCTAUrlButton
+          key={id}
+          body={this.text}
+          header={this.header}
+          headerType={this.headerType}
+          footer={this.footer}
+          displayText={this.displayText}
+          url={this.url}
+        />
+      )
+    }
+
     return (
       <WhatsappCTAUrlButton
         key={id}
         body={this.text}
-        header={this.header}
         footer={this.footer}
         displayText={this.displayText}
         url={this.url}