+++ packages/botonic-plugin-flow-builder/CHANGELOG.md
@@ -10,9 +10,11 @@ All notable changes to Botonic will be documented in this file.
     Click to see more.
   </summary>
   
-## [0.37.x] - 2025-mm-dd
+## [0.37.1] - 2025-mm-dd
 
 ### Added
+
+-[PR-3074](https://github.com/hubtype/botonic/pull/3074/files): Use Whatsapp CTA URL with image, video or document in header.
 
 ### Changed